+++ Comp2001CW2/Controllers/LanguagesController.cs
@@ -34,15 +34,24 @@ namespace Comp2001CW2.Controllers
         [HttpPost("CreateLanguage")]
         public IActionResult CreateLanguage(string languageID, string languageName)
         {
-            try
+            int? adminRights = HttpContext.Session.GetInt32("AdminRights");
+
+            if (adminRights != 1)
             {
-                _dbContext.CreateLanguage(languageID, languageName);
-                _dbContext.SaveChanges();
-                return Ok("Language created successfully.");
+                return BadRequest("Administrative rights required to create languages");
             }
-            catch (Exception ex)
+            else
             {
-                return StatusCode(500, $"Internal Server Error: {ex.Message}");
+                try
+                {
+                    _dbContext.CreateLanguage(languageID, languageName);
+                    _dbContext.SaveChanges();
+                    return Ok("Language created successfully.");
+                }
+                catch (Exception ex)
+                {
+                    return StatusCode(500, $"Internal Server Error: {ex.Message}");
+                }
             }
         }
 
@@ -64,15 +73,24 @@ namespace Comp2001CW2.Controllers
         [HttpDelete("DeleteLanguage")]
         public IActionResult DeleteLanguage(string languageID)
         {
-            try
+            int? adminRights = HttpContext.Session.GetInt32("AdminRights");
+
+            if (adminRights != 1)
             {
-                _dbContext.DeleteLanguage(languageID);
-                _dbContext.SaveChanges();
-                return Ok("Language deleted successfully.");
+                return BadRequest("Administrative rights required to delete languages");
             }
-            catch (Exception ex)
+            else
             {
-                return StatusCode(500, $"Internal Server Error: {ex.Message}");
+                try
+                {
+                    _dbContext.DeleteLanguage(languageID);
+                    _dbContext.SaveChanges();
+                    return Ok("Language deleted successfully.");
+                }
+                catch (Exception ex)
+                {
+                    return StatusCode(500, $"Internal Server Error: {ex.Message}");
+                }
             }
         }
